Исправляет ошибки в функциях , удаляет лишнии файлы
