+++ js/data.js
@@ -1,3 +1,4 @@
+import {getRandomInteger, getRandomElementFromArray, createRandomIdFromRangeGenerator} from './util.js';
 const NAMES =
 ['Артём',
   'Мария',
@@ -25,5 +26,31 @@ const DESCRIPTIONS = [
 ];
 
 const SIMILAR_PHOTO_COUNT = 25;
+const getRandomId = createRandomIdFromRangeGenerator(1, SIMILAR_PHOTO_COUNT);
+const getRandomUrl = createRandomIdFromRangeGenerator(1, SIMILAR_PHOTO_COUNT);
+const getRandomCommentId = createRandomIdFromRangeGenerator(1, SIMILAR_PHOTO_COUNT);
 
-export {NAMES, MESSAGES, DESCRIPTIONS, SIMILAR_PHOTO_COUNT};
+/**
+ * Функция создает массив случайных комментариев
+ * @returns {object} - объект со свойствами: id, avatar, message, name
+ */
+const createComments = () => ({
+  id: getRandomCommentId(),
+  avatar:`img/avatar-${getRandomInteger(1,6)}.svg`,
+  message: getRandomElementFromArray(MESSAGES),
+  name: getRandomElementFromArray(NAMES)
+});
+
+/**
+ * Функция создает массив случайных фотографий
+ * @returns {object} - объект со свойствами: id, url, description, likes, comments
+ */
+const createPhotos = () => ({
+  id: getRandomId(),
+  url: `photos/${getRandomUrl()}.jpg`,
+  description: getRandomElementFromArray(DESCRIPTIONS),
+  likes: getRandomInteger(15,200),
+  comments: Array.from({length: getRandomInteger(1,10)}, createComments)
+});
+
+export {SIMILAR_PHOTO_COUNT, createComments, createPhotos};
